+++ app/Http/Controllers/TodoListController.php
@@ -3,7 +3,10 @@
 
 namespace App\Http\Controllers;
 
+use App\Relation;
+use App\Schedule;
 use App\Todo;
+use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
@@ -13,16 +16,29 @@ class TodoListController extends Controller
 {
     public function index()
     {
-        $data = Todo::all();
+        $user  = Auth::user();
 
-        $data = json_decode($data, TRUE);
+        $data = DB::table('todos')
+                ->where('userId', '=', $user->id)
+                ->get()
+                ->toArray();
+
+//        $data = Todo::all();
+//
+//        $data = json_decode($data, TRUE);
+
+
+
+//        dd($data);
 
         return view('todoList', compact('data'));
     }
 
     public function create(Request $request)
     {
-        $data= $request->all();
+        $data = $request->all();
+
+//        dd($data);
         $data['userId'] = Auth::id();
         $data['done'] = 0;
 
@@ -32,15 +48,48 @@ class TodoListController extends Controller
             // todo komunikat za długi opis zadania
         }
 
-        Todo::create($data);
+        $todo = Todo::create($data);
+
+        if(isset($data['deadline']))
+        {
+            $schedule = new Schedule();
+
+            $schedule->startDate    = $data['deadline'];
+            $schedule->endDate      = $data['deadline'];
+            $schedule->name         = $data['name'];
+            $schedule->description  = $data['description'];
+            $schedule->userId       = $data['userId'];
+            $schedule->done         = false;
+
+            $schedule->save();
+
+            $relation = new Relation();
+
+            $relation->scheduleId   = $schedule->id;
+            $relation->todoId       = $todo->id;
+            $relation->userId       = $data['userId'];
+
+            $relation->save();
+        }
         return redirect()->route('todo');
     }
 
     public function markAsDone(Request $request)
     {
+        $user = Auth::user();
         $postData = $request->all();
 
         DB::unprepared("UPDATE todos SET done = CASE WHEN done = true THEN false ELSE true END where id = " . $postData['id']);
+
+        $relation = DB::table('relations')
+            ->where('scheduleId', '=', $postData['id'])
+            ->where('userId', '=', $user->id)
+            ->first();
+
+        if(isset($relation))
+        {
+            DB::unprepared("UPDATE schedules SET done = CASE WHEN done = true THEN false ELSE true END where id = " . $relation->scheduleId);
+        }
 
         return $this->index();
     }
@@ -49,10 +98,26 @@ class TodoListController extends Controller
     {
         $postData = $request->all();
 
+        $user = Auth::user();
+
         $task = Todo::findOrfail($postData['id']);
 
         $task->delete();
 
+        $relation = DB::table('relations')
+            ->where('todoId', '=', $postData['id'])
+            ->where('userId', '=', $user->id)
+            ->first();
+
+        if(isset($relation))
+        {
+            $rel = DB::table('relations')
+                ->where('relId', '=',  $relation->relId)
+                ->delete();
+
+            $task = Schedule::findOrfail($relation->scheduleId);
+            $task->delete();
+        }
         return redirect()->to('todo');
     }
 }
